adding main

+++ cpp/samples.h
@@ -1,7 +1,7 @@
 #ifndef SAMPLES_H
 #define SAMPLES_H
 
-// Returns sqaure area (for length size n).  For negative n, sqaure area is defined to be 0.
-int get_square_area(int n);
+// Returns square area (for length size n).  For negative n, square area is defined to be 0.
+float get_square_area(float n);
 
 #endif //SAMPLES_H
